# Backend_test_case/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "eigen",
    "contact": {},
    "version": "1.0",
    "description": "API Documentation Backend Test Case"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/pengembalian": {
      "post": {
        "tags": ["Book"],
        "summary": "pengembalian",
        "operationId": "pengembalian",
        "parameters": [],
        "requestBody": {
          "description": "pengembalian buku untuk body tgl_pengembalian bisa di kosongkan yang berarti menggunakan waktu saat ini",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/pengembalianrequest"
                  },
                  {
                    "example": {
                      "member": "M001",
                      "buku": "JK-45",
                      "tgl_pengembalian": "2023-05-20T00:00:00Z"
                    }
                  }
                ]
              },
              "example": {
                "member": "M001",
                "buku": "JK-45",
                "tgl_pengembalian": "2023-05-20T00:00:00Z"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Access-Control-Allow-Origin": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "*"
                  }
                }
              },
              "Content-Security-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "default-src 'self';base-uri 'self';font-src 'self' https: data:;form-action 'self';frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                  }
                }
              },
              "Cross-Origin-Embedder-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "require-corp"
                  }
                }
              },
              "Cross-Origin-Opener-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "same-origin"
                  }
                }
              },
              "Cross-Origin-Resource-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "same-origin"
                  }
                }
              },
              "X-DNS-Prefetch-Control": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "off"
                  }
                }
              },
              "X-Frame-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "SAMEORIGIN"
                  }
                }
              },
              "Strict-Transport-Security": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "max-age=15552000; includeSubDomains"
                  }
                }
              },
              "X-Download-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "noopen"
                  }
                }
              },
              "X-Content-Type-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "nosniff"
                  }
                }
              },
              "Origin-Agent-Cluster": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "?1"
                  }
                }
              },
              "X-Permitted-Cross-Domain-Policies": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "none"
                  }
                }
              },
              "Referrer-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "no-referrer"
                  }
                }
              },
              "X-XSS-Protection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "0"
                  }
                }
              },
              "Content-Length": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "32"
                  }
                }
              },
              "ETag": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"20-QUIAso4Veffqx3wyIE2g2c5VjV8\""
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Tue, 02 May 2023 05:14:42 GMT"
                  }
                }
              },
              "Connection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              },
              "Keep-Alive": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "timeout=5"
                  }
                }
              }
            },
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/success1"
                    },
                    {
                      "example": {
                        "data": "pengembalian berhasil"
                      }
                    }
                  ]
                },
                "example": {
                  "data": "pengembalian berhasil"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Access-Control-Allow-Origin": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "*"
                  }
                }
              },
              "Content-Security-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "default-src 'self';base-uri 'self';font-src 'self' https: data:;form-action 'self';frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                  }
                }
              },
              "Cross-Origin-Embedder-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "require-corp"
                  }
                }
              },
              "Cross-Origin-Opener-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "same-origin"
                  }
                }
              },
              "Cross-Origin-Resource-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "same-origin"
                  }
                }
              },
              "X-DNS-Prefetch-Control": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "off"
                  }
                }
              },
              "X-Frame-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "SAMEORIGIN"
                  }
                }
              },
              "Strict-Transport-Security": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "max-age=15552000; includeSubDomains"
                  }
                }
              },
              "X-Download-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "noopen"
                  }
                }
              },
              "X-Content-Type-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "nosniff"
                  }
                }
              },
              "Origin-Agent-Cluster": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "?1"
                  }
                }
              },
              "X-Permitted-Cross-Domain-Policies": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "none"
                  }
                }
              },
              "Referrer-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "no-referrer"
                  }
                }
              },
              "X-XSS-Protection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "0"
                  }
                }
              },
              "Content-Length": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "61"
                  }
                }
              },
              "ETag": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"3d-QavFhgsr+2dCu2Vdv8MdSa5Q9Ro\""
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Tue, 02 May 2023 05:10:16 GMT"
                  }
                }
              },
              "Connection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              },
              "Keep-Alive": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "timeout=5"
                  }
                }
              }
            },
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/bukutidaktersedia1"
                    },
                    {
                      "example": {
                        "data": {
                          "httpStatusCode": 404,
                          "error": "kode buku tidak ada"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "data": {
                    "httpStatusCode": 404,
                    "error": "kode buku tidak ada"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "headers": {
              "Access-Control-Allow-Origin": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "*"
                  }
                }
              },
              "Content-Security-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "default-src 'self';base-uri 'self';font-src 'self' https: data:;form-action 'self';frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
                  }
                }
              },
              "Cross-Origin-Embedder-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "require-corp"
                  }
                }
              },
              "Cross-Origin-Opener-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "same-origin"
                  }
                }
              },
              "Cross-Origin-Resource-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "same-origin"
                  }
                }
              },
              "X-DNS-Prefetch-Control": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "off"
                  }
                }
              },
              "X-Frame-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "SAMEORIGIN"
                  }
                }
              },
              "Strict-Transport-Security": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "max-age=15552000; includeSubDomains"
                  }
                }
              },
              "X-Download-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "noopen"
                  }
                }
              },
              "X-Content-Type-Options": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "nosniff"
                  }
                }
              },
              "Origin-Agent-Cluster": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "?1"
                  }
                }
              },
              "X-Permitted-Cross-Domain-Policies": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "none"
                  }
                }
              },
              "Referrer-Policy": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "no-referrer"
                  }
                }
              },
              "X-XSS-Protection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "0"
                  }
                }
              },
              "Content-Length": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "60"
                  }
                }
              },
              "ETag": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"3c-qIq40DPM0ezcbaoxj8iGGEccQ48\""
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Tue, 02 May 2023 05:12:02 GMT"
                  }
                }
              },
              "Connection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              },
              "Keep-Alive": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "timeout=5"
                  }
                }
              }
            },
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/invalidbody1"
                    },
                    {
                      "example": {
                        "data": {
                          "httpStatusCode": 422,
                          "error": "member is required"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "data": {
                    "httpStatusCode": 422,
                    "error": "member is required"
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "Data": {
        "title": "Error",
        "required": ["httpStatusCode", "error"],
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          }
        },
        "example": {
          "httpStatusCode": 500,
          "error": "error sistem"
        }
      },
...
      "pengembalianrequest": {
        "title": "pengembalianrequest",
        "required": ["member", "buku", "tgl_pengembalian"],
        "type": "object",
        "properties": {
          "member": {
            "type": "string"
          },
          "buku": {
            "type": "string"
          },
          "tgl_pengembalian": {
            "type": "string"
          }
        },
        "example": {
          "member": "M001",
          "buku": "JK-45",
          "tgl_pengembalian": "2023-05-20T00:00:00Z"
        }
      },
      "success1": {
        "title": "success1",
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "type": "string"
          }
        },
        "example": {
          "data": "pengembalian berhasil"
        }
      },
      "bukutidaktersedia1": {
        "title": "bukutidaktersedia1",
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Data"
          }
        },
        "example": {
          "data": {
            "httpStatusCode": 404,
            "error": "kode buku tidak ada"
          }
        }
      },
...
      "invalidbody1": {
        "title": "invalidbody1",
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Data"
          }
        },
        "example": {
          "data": {
            "httpStatusCode": 422,
            "error": "member is required"
          }
        }
      },
...
    }
  },
...
}